+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - New #61: Add `setStickerSetThumbnail` method.
 - New #62: Add `setCustomEmojiStickerSetThumbnail` method.
 - New #63: Add `setChatStickerSet` and `deleteChatStickerSet` methods.
+- New #64: Add `getForumTopicIconStickers` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.